requirements.txt
# Core
pydantic>=2.7,<3
PyYAML>=6.0.1,<7
pandas>=2.2,<3
pyarrow>=16,<17
numpy>=1.26,<3
tqdm>=4.66,<5

# NLP / ML
transformers>=4.42,<5
sentence-transformers>=3.0,<4
torch>=2.2,<3

# Retrieval
faiss-cpu>=1.8.0
rank-bm25>=0.2.2

# Parsing
unstructured>=0.15,<0.16
pymupdf>=1.24,<2
pdfplumber>=0.11,<0.12
python-docx>=1.1,<2
python-pptx>=0.6.23
nltk>=3.9,<4

# CLI
click>=8.1,<9
requests>=2.32,<3
streamlit>=1.36,<2
scikit-learn>=1.5,<2

# Testing
pytest>=7.4,<8

